updated StripePaymentGateway added generatePaymentLink

+++ src/main/java/devx/arjun/PaymentService/dto/PaymentLinkReqDTO.java
@@ -3,10 +3,10 @@ package devx.arjun.PaymentService.dto;
 public class PaymentLinkReqDTO {
     private String orderId;
     private String userId;
-    private double amount;
+    private long amount;
 
 
-    public PaymentLinkReqDTO(String orderId, String userId, double amount) {
+    public PaymentLinkReqDTO(String orderId, String userId, long amount) {
         this.orderId = orderId;
         this.userId = userId;
         this.amount = amount;
@@ -28,11 +28,11 @@ public class PaymentLinkReqDTO {
         this.userId = userId;
     }
 
-    public double getAmount() {
+    public long getAmount() {
         return amount;
     }
 
-    public void setAmount(double amount) {
+    public void setAmount(long amount) {
         this.amount = amount;
     }
 }
